+++ CHANGELOG.rst
@@ -15,9 +15,6 @@ dev
 - Added `get_dependencies` method to _Framework, to enable finding all parameters that a quantity depends on.
 
 **Bugfixes**
-v3.0.4 [23rd April 2019]
-------------------------
-**Bugfixes**
 
 - When using camb for the transfer function, some cosmologies would lead to a segfault (i.e. when Ob0 or Tcmb0 are not
   set explicitly). This now raises a helpful error.
@@ -27,6 +24,10 @@ v3.0.4 [23rd April 2019]
 - Removed logging, which was redundant.
 - Moved from nose to pytest
 - Significant overhaul of package structure to more modularised form.
+
+v3.0.4 [23rd April 2019]
+------------------------
+**Bugfixes**
 
 - Fixed bug which made CAMB transfer function not pickleable or copyable.
   NOTE: the fix is not perfect -- if you provide a custom CAMBparams class
